# resources/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API REST para Gestão de Clientes",
    "version": "1.0.0",
    "description": "Documentação completa da API de CRM simplificado com autenticação JWT e dashboard."
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "username": { "type": "string" },
          "password": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "vendedor"] },
          "status": { "type": "string", "enum": ["ativo", "inativo"] }
        }
      },
      "Client": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nome": { "type": "string" },
          "email": { "type": "string" },
          "telefone": { "type": "string" },
          "empresa": { "type": "string" },
          "endereco": {
            "type": "object",
            "properties": {
              "logradouro": { "type": "string" },
              "numero": { "type": "string" },
              "complemento": { "type": "string" },
              "bairro": { "type": "string" },
              "cidade": { "type": "string" },
              "estado": { "type": "string" },
              "cep": { "type": "string" }
            }
          },
          "observacoes": { "type": "string" },
          "status": { "type": "string", "enum": ["Aberto", "Fechado", "Perdido"] }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "statusCode": { "type": "integer" }
        }
      }
    }
  },
  "paths": {
    "/login": {
      "post": {
        "summary": "Login do usuário",
        "description": "Gera um token JWT para acesso às rotas protegidas.",
        "tags": ["Autenticação"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "example": {
                  "username": "admin",
                  "password": "admin123"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Operação realizada com sucesso.",
                  "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Usuário não autenticado ou token inválido.",
                  "statusCode": 401
                }
              }
            }
          }
        }
      }
    },
    "/usuarios": {
      "get": {
        "summary": "Lista usuários",
        "description": "Endpoint administrativo para listar usuários. Apenas perfis admin podem acessar.",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Lista de usuários retornada com sucesso.",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "username": "admin",
                    "role": "admin",
                    "status": "ativo"
                  }
                ]
              }
            }
          },
          "403": {
            "description": "Você não possui permissão para acessar este recurso.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Você não possui permissão para acessar este recurso.",
                  "statusCode": 403
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria usuário",
        "description": "Cria um novo usuário com perfil admin ou vendedor. A criação deve ocorrer por conta administrativa.",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário cadastrado com sucesso.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Usuário cadastrado com sucesso.",
                  "user": {
                    "id": 2,
                    "username": "vendedor1",
                    "role": "vendedor",
                    "status": "ativo"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/usuarios/{id}": {
      "get": {
        "summary": "Consulta usuário por identificador",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Usuário retornado." }
        }
      },
      "put": {
        "summary": "Atualiza usuário",
        "description": "Atualiza status, role ou username do usuário. Normalmente por perfil admin.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Usuário atualizado com sucesso." }
        }
      },
      "delete": {
        "summary": "Remove usuário",
        "description": "Remoção lógica do usuário. Recomendado para uma conta administrativa.",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Usuário removido com sucesso." }
        }
      }
    },
    "/clientes": {
      "get": {
        "summary": "Lista clientes",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "nome", "in": "query", "schema": { "type": "string" } },
          { "name": "email", "in": "query", "schema": { "type": "string" } },
          { "name": "telefone", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de clientes retornada com sucesso.",
            "content": {
              "application/json": {
                "example": [
                  {
                    "id": 1,
                    "nome": "Maria Silva",
                    "email": "maria@email.com",
                    "telefone": "11999999999",
                    "empresa": "Acme",
                    "endereco": {
                      "logradouro": "Rua A",
                      "numero": "123",
                      "complemento": "Bloco 1",
                      "bairro": "Centro",
                      "cidade": "São Paulo",
                      "estado": "SP",
                      "cep": "01000-000"
                    },
                    "observacoes": "Cliente importante",
                    "status": "Aberto"
                  }
                ]
              }
            }
          }
        }
      },
      "post": {
        "summary": "Cria cliente",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Client"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Cliente cadastrado com sucesso.",
            "content": {
              "application/json": {
                "example": {
                  "message": "Cliente cadastrado com sucesso.",
                  "client": {
                    "id": 1,
                    "nome": "Maria Silva",
                    "email": "maria@email.com",
                    "status": "Aberto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/clientes/{id}": {
      "get": {
        "summary": "Consulta um cliente por identificador",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Cliente retornado." }
        }
      },
      "put": {
        "summary": "Atualiza um cliente",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Cliente atualizado com sucesso." }
        }
      },
      "delete": {
        "summary": "Exclui um cliente",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Cliente excluído com sucesso." }
        }
      }
    },
    "/dashboard": {
      "get": {
        "summary": "Dashboard de indicadores",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Indicadores do dashboard.",
            "content": {
              "application/json": {
                "example": {
                  "totalClientes": 50,
                  "abertos": 18,
                  "fechados": 24,
                  "perdidos": 8
                }
              }
            }
          }
        }
      }
    }
  }
}